requirements.txt
PySide6
Pillow
piexif
face_recognition
setuptools
scikit-learn
mypy
numpy
opencv-python-headless
